+++ sdk/core/core-tracing/src/tracer.ts
@@ -1,48 +1,7 @@
-import { createTracingContext } from ".";
+import { Tracer, TracerCreateSpanOptions, TracingSpan, TracingContext } from "./interfaces";
+import { createTracingContext } from "./tracingContext";
 
-export interface TracerCreateSpanOptions {
-  context?: TracingContext;
-}
-export interface Tracer {
-  startSpan(
-    name: string,
-    options: TracerCreateSpanOptions
-  ): { span: TracingSpan; tracingContext: TracingContext };
-  // withTrace<
-  //   Callback extends (
-  //     context: TracingContext,
-  //     span: Omit<TracingSpan, "end">
-  //   ) => ReturnType<Callback>
-  // >(
-  //   name: string,
-  //   callback: Callback,
-  //   options: TracingSpanOptions,
-  //   callbackThis?: ThisParameterType<Callback>
-  // ): Promise<ReturnType<Callback>>;
-  withContext<Callback extends (args: Parameters<Callback>) => ReturnType<Callback>>(
-    callback: Callback,
-    options: TracerCreateSpanOptions,
-    callbackThis?: ThisParameterType<Callback>,
-    ...callbackArgs: Parameters<Callback>
-  ): ReturnType<Callback>;
-}
-
-export type SpanStatus =
-  | {
-      status: "success";
-    }
-  | {
-      status: "error";
-      error: Error | string;
-    };
-
-export interface TracingSpan {
-  setStatus(status: SpanStatus): void;
-  setAttribute(name: string, value: unknown): void;
-  end(): void;
-  unwrap(): unknown;
-}
-
+/** @internal */
 export class NoOpTracer implements Tracer {
   startSpan(
     _name?: string,
@@ -76,6 +35,7 @@ export class NoOpTracer implements Tracer {
     return callback.apply(callbackThis, callbackArgs);
   }
 }
+/** @internal */
 export class NoOpSpan implements TracingSpan {
   setStatus(): void {}
   setAttribute(): void {}
@@ -85,35 +45,7 @@ export class NoOpSpan implements TracingSpan {
   }
 }
 
-export interface TracingContext {
-  setValue(key: symbol, value: unknown): TracingContext;
-  getValue(key: symbol): unknown;
-  deleteValue(key: symbol): TracingContext;
-}
-
-export class TracingContextImpl implements TracingContext {
-  private _contextMap: Map<symbol, unknown>;
-  constructor(initialContext: Map<symbol, unknown>) {
-    this._contextMap = new Map<symbol, unknown>(initialContext);
-  }
-
-  setValue(key: symbol, value: unknown): TracingContext {
-    const newContextMap = new Map<symbol, unknown>(this._contextMap);
-    newContextMap.set(key, value);
-    return new TracingContextImpl(newContextMap);
-  }
-
-  getValue(key: symbol): unknown {
-    return this._contextMap.get(key);
-  }
-
-  deleteValue(key: symbol): TracingContext {
-    const newContextMap = new Map<symbol, unknown>(this._contextMap);
-    newContextMap.delete(key);
-    return new TracingContextImpl(newContextMap);
-  }
-}
-
+/** @internal */
 export let tracerImplementation: Tracer = new NoOpTracer();
 
 export function useTracer(tracer: Tracer): void {
