+++ sdk/core/core-tracing/src/utils/traceParentHeader.ts
@@ -1,64 +1,64 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
+// // Copyright (c) Microsoft Corporation.
+// // Licensed under the MIT license.
 
-import { SpanContext, TraceFlags } from "../interfaces";
+// import { SpanContext, TraceFlags } from "../interfaces";
 
-// TODO: where should this logic live?
-const VERSION = "00";
+// // TODO: where should this logic live?
+// const VERSION = "00";
 
-/**
- * Generates a `SpanContext` given a `traceparent` header value.
- * @param traceParent - Serialized span context data as a `traceparent` header value.
- * @returns The `SpanContext` generated from the `traceparent` value.
- */
-export function extractSpanContextFromTraceParentHeader(
-  traceParentHeader: string
-): SpanContext | undefined {
-  const parts = traceParentHeader.split("-");
+// /**
+//  * Generates a `SpanContext` given a `traceparent` header value.
+//  * @param traceParent - Serialized span context data as a `traceparent` header value.
+//  * @returns The `SpanContext` generated from the `traceparent` value.
+//  */
+// export function extractSpanContextFromTraceParentHeader(
+//   traceParentHeader: string
+// ): SpanContext | undefined {
+//   const parts = traceParentHeader.split("-");
 
-  if (parts.length !== 4) {
-    return;
-  }
+//   if (parts.length !== 4) {
+//     return;
+//   }
 
-  const [version, traceId, spanId, traceOptions] = parts;
+//   const [version, traceId, spanId, traceOptions] = parts;
 
-  if (version !== VERSION) {
-    return;
-  }
+//   if (version !== VERSION) {
+//     return;
+//   }
 
-  const traceFlags = parseInt(traceOptions, 16);
+//   const traceFlags = parseInt(traceOptions, 16);
 
-  const spanContext: SpanContext = {
-    spanId,
-    traceId,
-    traceFlags
-  };
+//   const spanContext: SpanContext = {
+//     spanId,
+//     traceId,
+//     traceFlags
+//   };
 
-  return spanContext;
-}
+//   return spanContext;
+// }
 
-/**
- * Generates a `traceparent` value given a span context.
- * @param spanContext - Contains context for a specific span.
- * @returns The `spanContext` represented as a `traceparent` value.
- */
-export function getTraceParentHeader(spanContext: SpanContext): string | undefined {
-  const missingFields: string[] = [];
-  if (!spanContext.traceId) {
-    missingFields.push("traceId");
-  }
-  if (!spanContext.spanId) {
-    missingFields.push("spanId");
-  }
+// /**
+//  * Generates a `traceparent` value given a span context.
+//  * @param spanContext - Contains context for a specific span.
+//  * @returns The `spanContext` represented as a `traceparent` value.
+//  */
+// export function getTraceParentHeader(spanContext: SpanContext): string | undefined {
+//   const missingFields: string[] = [];
+//   if (!spanContext.traceId) {
+//     missingFields.push("traceId");
+//   }
+//   if (!spanContext.spanId) {
+//     missingFields.push("spanId");
+//   }
 
-  if (missingFields.length) {
-    return;
-  }
+//   if (missingFields.length) {
+//     return;
+//   }
 
-  const flags = spanContext.traceFlags || TraceFlags.NONE;
-  const hexFlags = flags.toString(16);
-  const traceFlags = hexFlags.length === 1 ? `0${hexFlags}` : hexFlags;
+//   const flags = spanContext.traceFlags || TraceFlags.NONE;
+//   const hexFlags = flags.toString(16);
+//   const traceFlags = hexFlags.length === 1 ? `0${hexFlags}` : hexFlags;
 
-  // https://www.w3.org/TR/trace-context/#traceparent-header-field-values
-  return `${VERSION}-${spanContext.traceId}-${spanContext.spanId}-${traceFlags}`;
-}
+//   // https://www.w3.org/TR/trace-context/#traceparent-header-field-values
+//   return `${VERSION}-${spanContext.traceId}-${spanContext.spanId}-${traceFlags}`;
+// }
